+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Fixes
 
 - Fixes the styling for the `Image` component in the `Hero` "secondary" variant.
+- Adds `Tooltip` component.
 
 ## 1.0.8 (August 11, 2022)