+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Adds
 
 - Adds `Tooltip` component.
+- Adds the `TagSet` component.
 
 ### Fixes